+++ os/linux/rt_linux.c
@@ -2253,7 +2253,7 @@ BOOLEAN CFG80211_SupBandInit(
 	}
 
 	CFG80211DBG(RT_DEBUG_ERROR, ("80211> Number of channel = %d\n",
-				CFG80211_NUM_OF_CHAN_5GHZ));
+				(int) CFG80211_NUM_OF_CHAN_5GHZ));
 
 	if (pRates == NULL)
 	{
